Fix double-absolute SpaceDock URLs
