+++ Netkan/CKAN-netkan.csproj
@@ -97,6 +97,7 @@
     <Compile Include="Sources\Jenkins\JenkinsOptions.cs" />
     <Compile Include="Sources\Jenkins\JenkinsRef.cs" />
     <Compile Include="Sources\Spacedock\ISpaceDock.cs" />
+    <Compile Include="Sources\Spacedock\JsonConvertFromRelativeSdUri.cs" />
     <Compile Include="Sources\Spacedock\SDVersion.cs" />
     <Compile Include="Sources\Spacedock\SpaceDockApi.cs" />
     <Compile Include="Sources\Spacedock\SpacedockError.cs" />
